backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.4.0
redis==5.0.4
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
dnspython==2.6.1
httpx==0.27.0
python-dotenv==1.0.1
python-whois==0.9.4
reportlab==4.2.2
